Sending/Receiving messages now work
This commit is relatively larger as it involved fixing bugs in the server and client relating to endianness and UTF-8 before things could finally work. In particular, the server was switching endianness for each client, causing every even client to get sent the length of the message in host endian which breaks things on the client side as it had the incorrect length. UTF-8 caused bugs as the byte length != character count, and the incorrect version was being sent sometimes, causing utf-8 string validation to fail, and the message to therefore not appear in the text view.