Potentially fixed segfault upon connecting
After debugging with valgrind, I found that sometimes the client would pass from the setup window to the main window a non null-terminated string for the nickname. When in the main window the 'Connected as: nickname' string is formed, the append call under GString would run strlen and cause a segfault. This only happened sometimes, it shouldn't happen at all, as the string from the GtkEntry should be null-terminated, and this seems to be the case most of the time, but in this commit, a line of code to manually add in the null terminator has been added.